Organised the views a bit
