+++ src/main/java/app/CreateCommentUseCaseBuilder.java
@@ -10,7 +10,7 @@ import use_case.create_comment.interface_adapter.CreateCommentController;
 import use_case.create_comment.interface_adapter.CreateCommentPresenter;
 import view.CreateCommentView;
 import view.CreateCommentViewModel;
-import view.PostAndCommentsViewModel;
+import view.display_post.PostAndCommentsViewModel;
 
 public class CreateCommentUseCaseBuilder {
     private final PostAndCommentsViewModel postAndCommentsViewModel;
