+++ src/main/java/app/DisplayPostUseCaseFactory.java
@@ -5,8 +5,8 @@ import use_case.display_post.application_business_rules.DisplayPostInteractor;
 import use_case.display_post.application_business_rules.DisplayPostOutputBoundary;
 import use_case.display_post.interface_adapter.DisplayPostController;
 import use_case.display_post.interface_adapter.DisplayPostPresenter;
-import view.PostAndCommentsView;
-import view.PostAndCommentsViewModel;
+import view.display_post.PostAndCommentsView;
+import view.display_post.PostAndCommentsViewModel;
 
 public class DisplayPostUseCaseFactory {
     private DisplayPostUseCaseFactory() {}
