+++ src/main/java/use_case/create_comment/interface_adapter/CreateCommentPresenter.java
@@ -3,7 +3,7 @@ package use_case.create_comment.interface_adapter;
 import org.bson.types.ObjectId;
 import use_case.create_comment.application_business_rules.CreateCommentOutputBoundary;
 import view.CreateCommentViewModel;
-import view.PostAndCommentsViewModel;
+import view.display_post.PostAndCommentsViewModel;
 
 public class CreateCommentPresenter implements CreateCommentOutputBoundary {
     private final PostAndCommentsViewModel postAndCommentsViewModel;
