+++ src/main/java/use_case/display_post/interface_adapter/DisplayPostPresenter.java
@@ -2,7 +2,7 @@ package use_case.display_post.interface_adapter;
 
 import use_case.display_post.application_business_rules.DisplayPostOutputBoundary;
 import use_case.display_post.application_business_rules.DisplayPostOutputData;
-import view.PostAndCommentsViewModel;
+import view.display_post.PostAndCommentsViewModel;
 
 public class DisplayPostPresenter implements DisplayPostOutputBoundary {
     private final PostAndCommentsViewModel displayCommentViewModel;
