+++ src/main/java/view/display_post/PostView.java
@@ -1,4 +1,4 @@
-package view;
+package view.display_post;
 
 import org.bson.types.ObjectId;
