+++ src/test/java/use_case/create_comment/CreateCommentUseCaseTest.java
@@ -9,7 +9,7 @@ import use_case.create_comment.application_business_rules.CreateCommentInteracto
 import use_case.create_comment.interface_adapter.CreateCommentController;
 import use_case.create_comment.interface_adapter.CreateCommentPresenter;
 import view.CreateCommentViewModel;
-import view.PostAndCommentsViewModel;
+import view.display_post.PostAndCommentsViewModel;
 
 public class CreateCommentUseCaseTest {
     CreateCommentViewModel createCommentViewModel;
